Added interface for LocationsDao
